Stubbed out changes to the entity editor to allow the creation and amendment of AI states to an entity.
refactored the entity class's methods to assign states from json so that few conditionals are used and to use the map of states in the AIStateMachine class.

added a map of states to the AIStateMachine to make referring to the states easier and faster. these are to be used by the Agent class when loading states from json as well.

added a behavious json file as a template for agents with different behaviours. a behaviour is a group of pre-set states.