singled out any entities named Carrot to test behaviour changes in the entity class.
implemented the LoadBehaviourFile in the Agent class so that when a file path is passed to this method, the corresponding behaviour json file will be loaded and parsed. this creates the appropriate states and assigns to the states container.

next step is to change the the type of state container from map to something that allows duplicates like a vector of pairs.